+++ yoda/src/main/java/edu/cmu/sv/dialog_state_tracking/DiscourseUnitUpdateInference.java
@@ -20,6 +20,6 @@ public interface DiscourseUnitUpdateInference {
     * be weighted by the assumed hypothesis' prior and collected to create the new DU
     *
     * */
-    public DiscourseUnit2 applyAll(DiscourseUnit2.DialogStateHypothesis currentState, Turn turn, float timeStamp);
+    public DiscourseUnit2 applyAll(DiscourseUnit2.DialogStateHypothesis currentState, Turn turn, long timeStamp);
 
 }
